% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig.R
\name{binomialDiffCalc}
\alias{binomialDiffCalc}
\title{Differential methylation analysis using binomial model}
\usage{
binomialDiffCalc(meth, comparison, min.per.group = c(3, 3))
}
\arguments{
\item{meth}{A \code{BSseq-class} object to calculate differential methylation statistics. See \code{methylSigReadData} for how to read in methylation data.}

\item{comparison}{The name of the column in \code{pData(meth)} to use for the comparisons.}

\item{min.per.group}{A vector with two numbers specifying the minimum number of samples required to perform the test for differential methylation. If it is a single number, both groups will use it as the minimum requried number of samples. Default is \code{c(3,3)}.}
}
\value{
\code{GRanges} object containing the differential methylation statistics and locations. \code{p.adjust} with \code{method="BH"} option is used for p-value correction.
}
\description{
This function calculates differential methylation statistics using a binomial-based approach. See `Warning' message below.
}
\details{
This function uses a binomial-based model to calculate differential methylation statistics. It is nearly identical to the \code{methylKit::calculateDiffMeth} function in the \code{methylKit} R package except that only the likelihood ratio test and \code{p.adjust()} with \code{method=``BH''} are used to calculate significance levels. It is significantly faster than \code{methylKit::calculateDiffMeth} function.
}
\section{Warning}{
 This function does not take into account the variability among samples in each group being compared.
}

\examples{
data(data, package = 'methylSig')

myDiff = binomialDiffCalc(meth = data, comparison = 'DR_vs_DS')

}
\seealso{
\code{\link{methylSigCalc}}
}
\keyword{differentialMethylation}
