% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig-data.R
\docType{data}
\name{bsseq_cov_without_strand}
\alias{bsseq_cov_without_strand}
\title{BSseq object read from destranded coverage files}
\format{A BSseq object}
\source{
data-raw/create_test_data.R
}
\usage{
bsseq_cov_without_strand
}
\description{
Data contains 6 methylation loci and 2 samples
}
\examples{
data(bsseq_cov_without_strand, package = 'methylSig')
}
\keyword{datasets}
