% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig-data.R
\docType{data}
\name{bsseq_destranded}
\alias{bsseq_destranded}
\title{BSseq object read from destranded coverage files}
\format{A BSseq object}
\source{
data-raw/02-create_bsseq_rda.R
}
\usage{
bsseq_destranded
}
\description{
Data contains 6 methylation loci and 2 samples
}
\examples{
data(bsseq_destranded, package = 'methylSig')
}
\keyword{datasets}
