% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig-data.R
\docType{data}
\name{bsseq_multichrom}
\alias{bsseq_multichrom}
\title{BSseq object with loci on multiple chromosomes}
\format{
A BSseq object
}
\source{
data-raw/02-create_bsseq_rda.R
}
\usage{
bsseq_multichrom
}
\description{
Data contains 4 methylation loci for 2 samples on 2 chromosomes
}
\examples{
data(bsseq_multichrom, package = 'methylSig')
}
\keyword{datasets}
