% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig-data.R
\docType{data}
\name{bsseq_stranded}
\alias{bsseq_stranded}
\title{BSseq object read from stranded coverage files}
\format{
A BSseq object
}
\source{
data-raw/02-create_bsseq_rda.R
}
\usage{
bsseq_stranded
}
\description{
Data contains 11 methylation loci and 2 samples
}
\examples{
data(bsseq_stranded, package = 'methylSig')
}
\keyword{datasets}
