% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig_data_doc.R
\docType{data}
\name{data}
\alias{data}
\title{Example methylation data}
\usage{
data(data, package = 'methylSig')
}
\description{
A \code{BSseq-class} object containing 6 samples total, with three in each group. Genome is hg19.
}
\keyword{datasets}
