% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig_data_doc.R
\docType{data}
\name{data_CT_SNPs_hg19}
\alias{CT_SNPs_hg19}
\alias{data_CT_SNPs_hg19}
\title{CpG Index for hg19}
\format{A GenomicRanges object of length 1,321,463}
\source{
\url{ftp://ftp.1000genomes.ebi.ac.uk//vol1/ftp/release/20130502/ALL.wgs.phase3_shapeit2_mvncall_integrated_v5a.20130502.sites.vcf.gz}
}
\usage{
data(CT_SNPs_hg19)
}
\description{
A GenomicsRanges object giving the coordinates (in hg19) of all C > T SNPs.
Start coordinates are 0-based and end coordinates are 1-based. Starting from
1000 Genomes Data we used \code{bcftools filter} with \code{-i 'AF[0]>0.05'}
to pull all sites with alternate frequency greater than 0.05. We then used
\code{grep -P '(C\tT)'} and \code{grep -P '(VT=SNP)'} to collect all C > T
SNPs.
}
\keyword{datasets}

