% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig_data_doc.R
\docType{data}
\name{meth}
\alias{meth}
\title{Example methylSigData object}
\usage{
data(sampleData)
}
\description{
\code{meth} is a \code{\link{methylSigData}} object containing 8 samples total.
There are 4 AML (acute myeloid leukemia) samples and 4 NBM (normal bone marrow)
samples. Data are in hg18 coordinates.
}
\keyword{datasets}

