% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbs_enrichment.R
\name{methylSig.tfbsEnrichTest}
\alias{methylSig.tfbsEnrichTest}
\title{Perform transcription factor enrichment test among differentially methylated cytosines or regions}
\usage{
methylSig.tfbsEnrichTest(myDiff, dmcList, tfbsInfo)
}
\arguments{
\item{myDiff}{\code{GRanges} object resulting from \code{methylSigCalc} that contains all CpG sites that are tested for differential methylation.}

\item{dmcList}{A \code{logical} of the same length as \code{myDiff} defining the DMCs or DMRs.}

\item{tfbsInfo}{A \code{GRanges} object of the genomic regions representing peaks. The \code{name} column should indicate which TF the peak is for.}
}
\value{
A \code{data.frame} whose \code{rownames} are inherited from the \code{name} column of the input BED file, and whose columns are:
\describe{
  \item{n_total_by_tf}{ The number of tested CpGs in a TFBS for a TF. }
  \item{n_dmc_by_tf}{ The number of DM CpGs in a TFBS for a TF. }
  \item{N_total}{ The total number of tested CpGs in a TFBS across all the TFs. }
  \item{N_dmc}{ The total number of DM CpGs in a TFBS across all the TFs. }
  \item{p_total}{ \code{n_total_by_tf} / \code{N_total}, used in the likelihood calculation. }
  \item{p_dmc}{ \code{n_dmc_by_tf} / \code{N_dmc}, used in the likelihood calculation. }
  \item{logLik}{ The log-likelihood based on the binomial distribution. }
  \item{pvalue}{ The p-value from the likelihood ratio test. }
}
}
\description{
This function tests for enriched transcription binding sites among differentially methylated sites or regions using a binomial test.
}
\details{
Likelihood ratio test is used based on the binomial distribution.
}
\examples{
utils::data(sample_data, package = 'methylSig')

dmcList = msig_cpgs$fdr < 0.05 & abs(msig_cpgs$meth.diff) > 25

methylSig.tfbsEnrichTest(myDiff = msig_cpgs, dmcList = dmcList, tfbsInfo = tfbs)

}
