% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_methylSig.R
\name{methylSig_weightFunc}
\alias{methylSig_weightFunc}
\title{Default methylSig Weight Function}
\usage{
methylSig_weightFunc(u)
}
\arguments{
\item{u}{A numeric between 0 and 1}
}
\value{
A \code{GRanges} object containing the following \code{mcols}:
}
\description{
The default weight function used by methylSigCalc
}
\examples{
methylSig_weightFunc(0.5)

}
