% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multifactor.R
\name{methylSigDSS}
\alias{methylSigDSS}
\title{Calculates differential methylation statistics under general experimental design}
\usage{
methylSigDSS(meth, design, formula, contrast, group.term, min.per.group = c(3,
  3))
}
\arguments{
\item{meth}{A \code{BSseq-class} object to calculate differential methylation statistics. See \code{methylSigReadData} for how to read in methylation data.}

\item{design}{A \code{data.frame} for experimental design. Should contain as many rows as there are columns (samples) in \code{meth}.}

\item{formula}{A formula for the linear model. It should refer to column names from \code{design}. NOTE: The intercept is included by default if omitted. One can omit the intercept with a formula such as \code{'~ 0 + group'}. For clarity, it helps to include the intercept explicitly as in \code{'~ 1 + group'}.}

\item{contrast}{A contrast matrix for hypothesis testing. The number of rows should match the number of columns \code{design}.}

\item{group.term}{A string indicating which term in \code{formula} contains group information on which to apply the \code{min.per.group} parameter. Currently assumes that this factor contains ONLY TWO LEVELS.}

\item{min.per.group}{A vector with two numbers specifying the minimum number of samples required to perform the test for differential methylation. If it is a single number, both groups will use it as the minimum requried number of samples. Default is \code{c(3,3)}.  NOTE: The ordering of this parameter with respect to the groups should be \code{c(reference, other)}, where \code{reference} refers to the reference level in the \code{design[, group.term]} factor.}
}
\value{
A \code{GRanges} object containing the following \code{mcols}:
\describe{
  \item{stat}{ The dispersion estimate. }
  \item{pvalue}{ The log likelihood ratio. }
  \item{fdr}{ Degrees of freedom used when \code{T.approx = TRUE}. }
  \item{mean_methylation_columns}{ Mean methylation for each factor in each column of design, when there are fewer than 5 factors in the factor. }
}
}
\description{
Calculates differential methylation statistics under general experimental design
}
\examples{
data(data, package = 'methylSig')

# Example with implicit intercept
design1 = data.frame(group = bsseq::pData(data)$DR_vs_DS)
contrast1 = matrix(c(0,1), ncol = 1)
result1 = methylSigDSS(
    meth = data,
    design = design1,
    formula = '~ group',
    contrast = contrast1,
    group.term = 'group',
    min.per.group=c(3,3))

# Example with no intercept and corresponding change in contrast
# NOTE: result2 is the same as result1
contrast2 = matrix(c(-1,1), ncol = 1)
result2 = methylSigDSS(
    meth = data,
    design = design1,
    formula = '~ 0 + group',
    contrast = contrast2,
    group.term = 'group',
    min.per.group=c(3,3))

# Example with subject pairing
design2 = data.frame(
    group = bsseq::pData(data)$DR_vs_DS,
    subject = factor(c(1,1,2,2,3,3)))
contrast2 = matrix(c(0,1,0,0), ncol = 1)
result2 = methylSigDSS(
    meth = data,
    design = design2,
    formula = '~ group + subject',
    contrast = contrast2,
    group.term = 'group',
    min.per.group=c(3,3))

}
\seealso{
\code{\link{methylSigReadData}}
}
\keyword{differentialMethylation}
