% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{methylSigPlotDiff}
\alias{methylSigPlotDiff}
\title{Data visualization function}
\usage{
methylSigPlotDiff(myAnnots, annotation_order = NULL, status_order = NULL,
  bin_width = 10, plot_title = "Methylation Differences",
  x_label = "Methylation Difference", y_label = "Density",
  legend_facet_label = "Meth. Diff. in Annotation",
  legend_cum_label = "Meth. Diff. Overall")
}
\arguments{
\item{myAnnots}{A \code{GRanges} object resulting from running \code{methylSigAnnotation} on the result of \code{methylSigCalc}.}

\item{annotation_order}{A character vector which orders and subsets the annotations for the plot.}

\item{status_order}{A character vector which orders and subsets the DM status.}

\item{bin_width}{A vector of two numbers (from, to) to specify the region to visualize on chromosome `chr'.}

\item{plot_title}{The plot label.}

\item{x_label}{The x-axis label.}

\item{y_label}{The y-axis label.}

\item{legend_facet_label}{Label explaining the meaning of the gray bars in the resulting histogram.}

\item{legend_cum_label}{Label explaining the meaning of the red outlined bars in the resulting histogram.}

\item{legend_title}{The legend label.}
}
\value{
A \code{ggplot2} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Generates data visualization plot of methylation data for a specified genomic interval.
}
\details{
This function offers a unique two-tiered visualization of the methylation data depending on the zoom level. For narrow regions (<1mbp) where at most 500 CpG sites have data reads, users can visualize sample-specific coverage levels and percent methylation at each site, together with group averages, significance levels and a number of genomic annotations.
}
\examples{
# Annotate the msig_cpgs results
data(data, package = 'methylSig')

# Use the genome of msig_cpgs and build annotations for CpG features
genome = GenomeInfoDb::genome(msig_cpgs)
annots = annotatr::build_annotations(genome = genome, annotations = paste(genome, c('cpgs'), sep='_'))

# Decide what counts as differentially methylated
dmcList = msig_cpgs$fdr < 0.05 & abs(msig_cpgs$meth.diff) > 25

# Annotate
myDiff_annotated = methylSigAnnotation(myDiff = msig_cpgs, dmcList = dmcList, annotations = annots)

# Set the order vectors
cpg_order = c('hg19_cpg_islands','hg19_cpg_shores','hg19_cpg_shelves','hg19_cpg_inter')
status_order = c('hyper','hypo','none')

methylSigPlotDiff(myAnnots = myDiff_annotated, annotation_order = cpg_order, status_order = status_order,
    bin_width = 10, plot_title = 'Meth. Diff. in CpG Annots.', x_label = 'DM Status', y_label = 'Proportion',
    legend_facet_label = 'Meth. Diff. in Annotation', legend_cum_label = 'Meth. Diff. Overall')

}
