% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig_data_doc.R
\docType{data}
\name{msig_tiles}
\alias{msig_tiles}
\title{Example tile-level methylSigCalc results}
\usage{
data(data, package = 'methylSig')
}
\description{
A \code{GRanges-class} object containing the results of \code{methylSigCalc} on \code{tiled_data}.
}
\keyword{datasets}
