% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig_data_doc.R
\docType{data}
\name{myDiffSigboth}
\alias{myDiffSigboth}
\title{Example methylSigDiff object}
\usage{
data(sampleData)
}
\description{
\code{mySigDiffboth} is a \code{\link{methylSigDiff}} object storing the result
of the site-wise differential methylation test using \code{\link{methylSigCalc}}.
}
\keyword{datasets}

