% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig-data.R
\docType{data}
\name{promoters_gr}
\alias{promoters_gr}
\title{GRanges object with collapsed promoters on chr21 and chr22}
\format{
A GRanges object
}
\source{
data-raw/02-create_bsseq_rda.R
}
\usage{
promoters_gr
}
\description{
Data contains 1466 promoters for use in the vignette
}
\examples{
data(promoters_gr, package = 'methylSig')
}
\keyword{datasets}
