% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{readBismarkData}
\alias{readBismarkData}
\title{Read output from bismark_methylation_extractor to make a 'methylSigData' object.}
\usage{
readBismarkData(bismarkCovFiles, cytosineCovFiles, sample.ids, assembly = NA,
  pipeline = NA, context = NA, resolution = "base", treatment,
  destranded = TRUE, maxCount = 500, minCount = 10, filterSNPs = FALSE,
  num.cores = 1, quiet = FALSE)
}
\arguments{
\item{bismarkCovFiles}{Vector of coverage files (.cov as of Bismark v0.13.0) from \code{bismark_methylation_extractor} with \code{--bedGraph} flag. Sample order should match that of \code{cytosineCovFiles}.}

\item{cytosineCovFiles}{Vector of cytosine report files from \code{bismark_methylation_extractor} with \code{--cytosine_report} flag. Sample order should match that of \code{bismarkCovFiles}.}

\item{sample.ids}{Vector of sample ids.}

\item{assembly}{Character string indicating the genome assembly, such as "hg18", "hg19", "mm9", or "mm10".}

\item{pipeline}{Character string indicating the pipepline name that generated the data, for example, "bismark".}

\item{context}{Methylation context string such "CpG","CpH",or "CHH".}

\item{resolution}{A string indicating whether the input data are  base-pair or regional resolution. Either "base" or "region" is allowed. Default is "base".}

\item{treatment}{A numeric vector contatining numbers to distinguish the group classification.}

\item{destranded}{A logical value indicating whether to destrand the reverse to forward strand. If TRUE, the reads from both will be combined. Default is TRUE.}

\item{maxCount}{A number indicating the maximum coverage count to be included.}

\item{minCount}{A number indicating the minimum coverage count to be included.}

\item{filterSNPs}{A logical value indicating whether or not to filter out C > T SNPs based on the 1000 Genomes Project.}

\item{num.cores}{Number of cores to be used in reading files. Default is 1.}

\item{quiet}{A logical value. If quiet=TRUE, then this function does not show progress information. Default is FALSE.}
}
\value{
A 'methylSigData' object.
}
\description{
This function takes the coverage and cytosine report files from the \code{bismark_methylation_extractor} (options \code{--bedGraph} and \code{--cytosine_report}) and constructs a table conforming to that expected by the 'methylSigData' class.
}
\seealso{
\code{\link{methylSigCalc}}
}

