% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig_data_doc.R
\docType{data}
\name{tfbs}
\alias{tfbs}
\title{Example TFBS sites}
\usage{
data(data, package = 'methylSig')
}
\description{
A \code{GRanges-class} object representing transcription factor binding sites. For use in \code{methylSig.tfbsEnrichTest}. Genome is hg19.
}
\keyword{datasets}
