% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylSig_data_doc.R
\docType{data}
\name{tiled_data}
\alias{tiled_data}
\title{Example tiled methylation data}
\usage{
data(data, package = 'methylSig')
}
\description{
A tiled version of the \code{BSseq-class} object called \code{data}. Tiles are 1000bp. Genome is hg19.
}
\keyword{datasets}
