% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{write.methylSigDiff}
\alias{write.methylSigDiff}
\title{Write a methylSigDiff object as a text file}
\usage{
write.methylSigDiff(object, ...)
}
\arguments{
\item{...}{Arguments inherited from \code{\link[utils]{write.table}}}

\item{Object}{a `methylSigDiff' object.}
}
\value{
None
}
\description{
This funciton write methylSigDiff as a text file. All options for write.table are avaiable for this function.
}
\examples{
data(methylSig)
write.methylSigDiff(myDiffSigboth, file="myResult.txt", row.names=FALSE, quote=FALSE, sep="\\t")
}

