% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{cpgAnnotation}
\alias{cpgAnnotation}
\title{CpG island annotation function}
\usage{
cpgAnnotation(cpgInfo, myDiff)
}
\arguments{
\item{cpgInfo}{A CpG island information object.}

\item{myDiff}{A `methylSigDiff' object to be used for annotation.}
}
\value{
A factor vector with the same size as `myDiff' that indicates whether the related sites in `myDiff' are in a CpG island, CpG shore, CpG shelf or interCGI region.
}
\description{
This function annotates the CpG sites or regions to CpG island, CpG shore, CpG shelf or interCGI regions.
}
\details{
This function generates an object to annotate CpG island information. The same genome assembly should be used for `myDiff' and `cpgInfo'. CpG shores are defined as the region outside CpG islands but within 2,000 bp of any CpG island. CpG shelves are defined as the region within 2,000 bp away from a CpG shore. When regions overlap, the priority is CpG island, then CpG shore. Regions >4,000 bp from a CpG island are deemed interCGI.
}
\examples{
data(methylSig)

cpgInfo = getCpGInfo(system.file("annotation", "cpgi.hg18.bed.txt",
                                 package = "methylSig"))

cpgAnn = cpgAnnotation(cpgInfo,myDiffSigboth)
}
\seealso{
\code{\link{cpgAnnotationPlot}} and \code{\link{getCpGInfo}}
}
\keyword{CpGAnnotation}

