% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cpgAnnotationPlot}
\alias{cpgAnnotationPlot}
\title{CpG island Annotation pie chart}
\usage{
cpgAnnotationPlot(listFrom, main = "Plot", color, noShelf = FALSE)
}
\arguments{
\item{listFrom}{An R object generated by \code{\link{cpgAnnotation}} function.}

\item{main}{An overall title for the plot. The default is "Plot".}

\item{color}{Colors used in the plot to show CpG island, CpG shore, CpG shelve and interCGI regions. If noShelf = TRUE, then only three colors are needed.}

\item{noShelf}{Whether CpG shelves should be included in the plot. If noShelf = TRUE, then CpG shelves are treated as part of interCGI region. Default is FALSE.}
}
\description{
This function generates a pie chart that shows the proportions of CpG sites annotated to a CpG island, CpG shore, CpG shelf or interCGI region.
}
\details{
CpG shore is defined as the region outside of a CpG island but within 2,000 bp of any CpG island. CpG shelf is defined as the region 2,000-4,000bp away from a CPG island. When overlaped, the priority is CpG island, then CpG shore.
}
\examples{
data(sampleData)

cpgInfo = getCpGInfo(system.file("annotation", "cpgi.hg18.bed.txt",
                       package = "methylSig"))

cpgAnn = cpgAnnotation(cpgInfo,myDiffSigboth)

cpgAnnotationPlot(cpgAnn,main="ALL")

}
\seealso{
\code{\link{cpgAnnotation}} and \code{\link{getCpGInfo}}
}
\keyword{CpGAnnotation}
