% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_dss_test.R
\name{diff_dss_test}
\alias{diff_dss_test}
\title{Calculates differential methylation statistics under general experimental design}
\usage{
diff_dss_test(diff_fit, contrast, methylation_group_column, methylation_groups)
}
\arguments{
\item{diff_fit}{a \code{BSseq-class} object to calculate differential methylation statistics.}

\item{contrast}{a contrast matrix for hypothesis testing. The number of rows should match the number of columns \code{design}.}

\item{methylation_group_column}{a thing.}

\item{methylation_groups}{a thing.}
}
\value{
A \code{GRanges} object containing the following \code{mcols}:
\describe{
  \item{stat}{ The test statistic }
  \item{pvalue}{ p-values }
  \item{fdr}{ Degrees of freedom used when \code{T.approx = TRUE}. }
}
}
\description{
This function is a wrapper for \code{DSS::DMLfit.multiFactor} and \code{DSS::DMLtest.multiFactor} with the added feature of reporting methylation rates alongside the test results via the \code{methylation_group_column} and \code{methylation_groups} parameters. See documentation below.
}
