% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{getCpGInfo}
\alias{getCpGInfo}
\title{Read CpG island information from a text file.}
\usage{
getCpGInfo(infile)
}
\arguments{
\item{infile}{A path to a BED file containing chromosome, start, and end positions for CpG islands.}
}
\value{
CpG island information database that can be used in function cpgAnnotation.
}
\description{
This function reads CpG island information from a text or bed file for future annotation.
}
\section{Warning}{
 The same genome assembly of data should be used in annotation. This function does not check genome assembly type or version.
}
\examples{
cpgInfo = getCpGInfo(system.file("annotation", "cpgi.hg18.bed.txt",
                      package = "methylSig"))
}
\seealso{
\code{\link{cpgAnnotationPlot}} and \code{\link{cpgAnnotation}}
}
\keyword{CpGAnnotation}

