% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{getRefgeneInfo}
\alias{getRefgeneInfo}
\title{Read RefGene model information.}
\usage{
refGeneInfo = getRefgeneInfo(system.file("annotation", "refGene.txt",
                            package = "methylSig"))
}
\arguments{
\item{infile}{A path to file that contains refGene information.}
}
\value{
RefGene information object that can be used in function \code{\link{refGeneAnnotation}}.
}
\description{
This function reads RefGene model information for future annotation.
}
\details{
This function reads RefGene informaiton for future annotation. Same genome assembly of data should be used in annotation. This function does not check genome assembly type nor version.
}
\seealso{
\code{\link{refGeneAnnotation}} and \code{\link{refGeneAnnotationPlot}}
}
\keyword{refGeneAnnotation}

