% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{getTFBSInfo}
\alias{getTFBSInfo}
\title{Read transcription factor binding site information}
\usage{
getTFBSInfo(infile)
}
\arguments{
\item{infile}{A path to BED4+2 file (\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}) containing transcription factor binding site information.}
}
\value{
An object containing transcription factor binding sites information is returned. It can be used in function \code{\link{methylSig.tfbsEnrichTest}}.
}
\description{
This function reads the transcription factor binding site information for future annotation.
}
\examples{
   tfbsInfo = getTFBSInfo(system.file("annotation", "tfbsUniform.txt",
                       package = "methylSig"))

}
\seealso{
\code{\link{methylSig.tfbsEnrichTest}}
}
\keyword{tfbsAnnotation}

