% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfbs_enrichment.R
\name{methylSig.tfbsEnrichTest}
\alias{methylSig.tfbsEnrichTest}
\title{Perform transcription factor enrichment test among differentially methylated cytosines or regions}
\usage{
methylSig.tfbsEnrichTest(myDiff, dmcList, tfbsInfo, plot = TRUE,
  max.plot.num = 10, main = "tfbs Enrichment")
}
\arguments{
\item{myDiff}{\code{methylSigDiff-class} object that contains all CpG sites that are tested for differential methylation.}

\item{dmcList}{A boolean vector of the same length as myDiff defining the DMCs or DMRs. A value of TRUE for dmcList[i] means that the ith CpG site in myDiff is a DMC.}

\item{tfbsInfo}{An object that contains transcription factor binding sites information.}

\item{plot}{If plot = TRUE, this function draws a plot. Otherwise no plot is given. Default is TRUE.}

\item{max.plot.num}{Maximum number of transcription factor binding sites on the plot. Default is 10.}

\item{main}{An overall title for the plot.}
}
\value{
p-values of enriched transcription factor binding sites.
}
\description{
This function tests for enriched transcription binding sites among differentially methylated sites or regions using a binomial test. It also has an option to generate a bar plot to show the significance levels for the enriched transcription factor binding sites.
}
\details{
Likelihood ratio test is used based on the binomial distribution.
}
\examples{
data(sampleData)

tfbsInfo = getTFBSInfo(system.file("annotation", "tfbsUniform.txt",
                        package = "methylSig"))

pvalue = methylSig.tfbsEnrichTest(myDiffSigboth,
         myDiffSigboth[,"qvalue"]<0.05 & abs(myDiffSigboth[,"meth.diff"])>25,
         tfbsInfo)

}
\seealso{
\code{\link{getTFBSInfo}}
}
\keyword{tfbsAnnotation}

