% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_methylSig.R
\name{methylSigCalc}
\alias{methylSigCalc}
\title{Calculates differential methylation statistics using a Beta-binomial approach.}
\usage{
methylSigCalc(meth, comparison = NA, dispersion = "both",
  local.info = FALSE, local.winsize = 200, min.per.group = c(3, 3),
  weightFunc = methylSig_weightFunc, T.approx = TRUE, num.cores = 1)
}
\arguments{
\item{meth}{A \code{BSseq-class} object to calculate differential methylation statistics. See \code{methylSigReadData} for how to read in methylation data.}

\item{comparison}{The name of the column in \code{pData(meth)} to use for the comparisons, with the correct reference level set.}

\item{dispersion}{One of \code{both}, or either group name. Indicates which set of samples to use to estimate the dispersion parameter. Default is \code{both}.}

\item{local.info}{A \code{logical} value indicating whether to use local information to improve mean and dispersion parameter estimations. Default is \code{FALSE}.}

\item{local.winsize}{An \code{integer} to specify the distance upstream and downstream of a location to include local information for the mean and dispersion parameter estimations. NOTE: An additional constraint is placed whereby a maximum of 5 loci upstream and downstream of the locus of interest are used. Default is \code{200}.}

\item{min.per.group}{A vector with two numbers specifying the minimum number of samples required to perform the test for differential methylation. If it is a single number, both groups will use it as the minimum requried number of samples. Default is \code{c(3,3)}. NOTE: The ordering of this parameter with respect to the groups should be \code{c(reference, other)}, where \code{reference} refers to the reference level in the \code{pData(meth)[, comparison]} factor.}

\item{weightFunc}{A weight kernel function. The input of this function is from -1 to 1. The default is the tri-weight kernel function defined as \code{function(u) = (1-u^2)^3}. Function value and range of parameter for weight function should be from 0 to 1.}

\item{T.approx}{A \code{logical} value indicating whether to use squared t approximation for the likelihood ratio statistics. Chi-square approximation (\code{T.approx = FALSE}) is recommended when the sample size is large.  Default is \code{TRUE}.}

\item{num.cores}{An integer denoting how many cores should be used for differential methylation calculations.}
}
\value{
A \code{GRanges} object containing the following \code{mcols}:
\describe{
  \item{phiCommonEst}{ The dispersion estimate. }
  \item{logLikRatio}{ The log likelihood ratio. }
  \item{df}{ Degrees of freedom used when \code{T.approx = TRUE}. }
  \item{muEstC_group1}{ Methylation estimate for group1. Groups correspond to the levels in the column used for the comparison in \code{pdata(meth)}. }
  \item{muEstC_group2}{ Methylation estimate for group2. }
  \item{meth.diff}{ The difference \code{muEstC_group2 - muEstC_group1}. }
  \item{hyper.direction}{ The group for which the CpG/region is hyper-methylated. Groups correspond to the levels in the column used for the comparison in \code{pdata(meth)}. }
  \item{pvalue}{ The p-value from the t-test (\code{T.approx = TRUE}) or the Chi-Square test (\code{T.approx = FALSE}). }
  \item{fdr}{ The Benjamini-Hochberg adjusted p-values using \code{p.adjust(method = 'BH')}. }
}
}
\description{
The function calculates differential methylation statistics between two groups of samples. This is the main function of the methylSig package, and the method most users should use to test for DMCs or DMRs. The function uses a Beta-binomial approach to calculate differential methylation statistics, accounting for variation among samples within each group.
}
\details{
The function calculates differential methylation statistics between two groups of samples. The function uses Beta-binomial approach to calculate differential methylation statistics, accounting for variation among samples within each group. Users who wish to tile their data and test for differentially methylated regions (DMRs) instead DMCs should first use the \code{\link{methylSigTile}} function before using this function.
}
\examples{
data(data, package = 'methylSig')

result = methylSigCalc(
    meth = data,
    comparison = 'DR_vs_DS',
    dispersion = 'both',
    local.info = FALSE,
    local.winsize = 200,
    min.per.group = c(3,3),
    weightFunc = methylSig_weightFunc,
    T.approx = TRUE,
    num.cores = 1)

}
\seealso{
\code{\link{methylSigPlot}}, \code{\link{methylSigReadData}}
}
\keyword{differentialMethylation}
