% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{methylSigData-class}
\alias{[}
\alias{[,methylSigData,ANY,ANY-method}
\alias{methylSigData}
\alias{methylSigData-class}
\alias{show,methylSigData-method}
\title{This class is for generating 'methylSigData' object.}
\usage{
\S4method{[}{methylSigData,ANY,ANY}(x, i, j, drop)

\S4method{show}{methylSigData}(object)
}
\arguments{
\item{i}{Start position}

\item{j}{End position}

\item{object}{A methylSigData object.}
}
\description{
Objects can be created by calls of the form \code{new("methylSigData", ...)}.

Subset methylSigData object based on a range

This method is for printing a methylSigData object.
}
\section{Slots}{

\describe{
\item{\code{data.ids}}{A numeric vector}

\item{\code{data.chr}}{A factor vector}

\item{\code{data.strand}}{A factor vector}

\item{\code{data.start}}{A numeric vector}

\item{\code{data.end}}{A numeric vector}

\item{\code{data.coverage}}{A matrix}

\item{\code{data.numCs}}{A matrix}

\item{\code{data.numTs}}{A matrix}

\item{\code{treatment}}{A numeric vector}

\item{\code{sample.ids}}{A character vector}

\item{\code{sample.filenames}}{A character vector}

\item{\code{destranded}}{A logical value}

\item{\code{resolution}}{A character value}

\item{\code{options}}{A character vector}
}}
\examples{
showClass("methylSigData")
show("methylSigData")
}
\keyword{classes}
\keyword{internal}
\keyword{methods}

