% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{methylSigDiff-class}
\alias{[,methylSigDiff,ANY,character-method}
\alias{[,methylSigDiff,ANY,missing-method}
\alias{methylSigDiff}
\alias{methylSigDiff-class}
\alias{show,methylSigDiff-method}
\title{This class is for generating 'methylSigDiff' object.}
\usage{
\S4method{[}{methylSigDiff,ANY,character}(x, i, j)

\S4method{[}{methylSigDiff,ANY,missing}(x, i)

\S4method{show}{methylSigDiff}(object)
}
\arguments{
\item{i}{Start position}

\item{j}{End position}

\item{object}{A methylSigDiff object.}

\item{i}{Start position}

\item{j}{End position}
}
\description{
Objects can be created by calls of the form \code{new("methylSigDiff", ...)}.

Subset methylSigDiff object

Subset methylSigDiff object

This method is for printing a methylSigDiff object.
}
\section{Slots}{

\describe{
\item{\code{data.ids}}{A numeric vector}

\item{\code{data.chr}}{A factor vector}

\item{\code{data.strand}}{A factor vector}

\item{\code{data.start}}{A numeric vector}

\item{\code{data.end}}{A numeric vector}

\item{\code{results}}{A matrix}

\item{\code{treatment}}{A numeric vector}

\item{\code{sample.ids}}{A character vector}

\item{\code{sample.filenames}}{A character vector}

\item{\code{destranded}}{A logical value}

\item{\code{resolution}}{A character value}

\item{\code{options}}{A character vector}

\item{\code{data.options}}{A character vector}
}}
\examples{
show("myDiff")
}
\keyword{classes}
\keyword{internal}
\keyword{methods}

