% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{methylSigPlotStatus}
\alias{methylSigPlotStatus}
\title{Plot DM status distribution in annotations}
\usage{
methylSigPlotStatus(myAnnots, annotation_order = NULL, status_order = NULL,
  position = "fill", plot_title = "DM Status",
  legend_title = "Annotations", x_label = "DM Status",
  y_label = "Proportion")
}
\arguments{
\item{myAnnots}{A \code{GRanges} object resulting from running \code{methylSigAnnotation} on the result of \code{methylSigCalc}.}

\item{annotation_order}{A character vector which orders and subsets the annotations for the plot.}

\item{status_order}{A character vector which orders and subsets the DM status.}

\item{position}{One of \code{fill}, \code{stack}, \code{dodge}. See \code{ggplot2} documentation for details.}

\item{plot_title}{The plot label.}

\item{legend_title}{The legend label.}

\item{x_label}{The x-axis label.}

\item{y_label}{The y-axis label.}
}
\value{
A \code{ggplot2} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Generates a bar plot showing the distribution of differential methylation status of loci in selected annotations.
}
\examples{
# Annotate the msig_cpgs results
data(data, package = 'methylSig')

# Use the genome of msig_cpgs and build annotations for CpG features
genome = GenomeInfoDb::genome(msig_cpgs)
annots = annotatr::build_annotations(genome = genome, annotations = paste(genome, c('cpgs'), sep='_'))

# Decide what counts as differentially methylated
dmcList = msig_cpgs$fdr < 0.05 & abs(msig_cpgs$meth.diff) > 25

# Annotate
myDiff_annotated = methylSigAnnotation(myDiff = msig_cpgs, dmcList = dmcList, annotations = annots)

# Set the order vectors
cpg_order = c('hg19_cpg_islands','hg19_cpg_shores','hg19_cpg_shelves','hg19_cpg_inter')
dm_order = c('DR','DS','No DM')

methylSigPlotStatus(myAnnots = myDiff_annotated, annotation_order = cpg_order, status_order = dm_order,
    position = 'fill', plot_title = 'DM Status in CpG Annots.', legend_title = 'Annotations',
    x_label = 'DM Status', y_label = 'Proportion')

}
