% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{methylSigTile}
\alias{methylSigTile}
\title{Obtain tiled methylation data in non-overlapping continuous windows.}
\usage{
methylSigTile(meth, tiles = NULL, win.size = 25)
}
\arguments{
\item{meth}{A \code{methylSigData-class} object used to tile data.}

\item{tiles}{A data.frame including columns with chr, start, end columns for predefined tiles. Those CpG sites not belonging any tile will be removed from tiled data. If \code{tiles} is not \code{NULL}, then \code{win.size} is ignored.}

\item{win.size}{An integer value indicating the desired window size in bps. Default is 25. If \code{tiles} is not \code{NULL}, then \code{win.size} is ignored.}
}
\value{
A \code{methylSigData-class} object.
}
\description{
This funciton tiles data within windows of a given width across genome. It gives total number of methylated cytosines and total number of reads (coverage) within each tiled window. This is used prior to the \code{\link{methylSigCalc}} function when the user prefers to conduct a tiled analysis instead of a base specific analysis for differential methylation. Tiling may provide higher power to detect significant differences, especially for experiments with low coverage.
}
\examples{
data(sampleData)
methTile = methylSigTile(meth)
}

