% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{methylSigTile}
\alias{methylSigTile}
\title{Obtain tiled methylation data in non-overlapping continuous windows.}
\usage{
methylSigTile(meth, tiles = NULL, win.size = 200)
}
\arguments{
\item{meth}{A \code{BSseq-class} object, as from \code{methylSigReadData}.}

\item{tiles}{One of \code{NULL}, a \code{data.frame}, or a \code{GRanges} object. Those CpG sites not belonging any tile will be removed from tiled data.}

\item{win.size}{An \code{integer} indicating the desired window size in bps. Default is 200. Used only when \code{tiles = NULL}.}
}
\value{
A \code{BSseq-class} object.
}
\description{
This function summarizes methylation data within tiles or user-specified regions. For all CpGs within an intersecting genomic region, the coverage and methylation reads are summed. This is used prior to the \code{\link{methylSigCalc}} function when the user prefers to conduct a tiled analysis instead of a base specific analysis for differential methylation. Tiling may provide higher power to detect significant differences, especially for experiments with low coverage.
}
\examples{
data(data, package = 'methylSig')
methTile = methylSigTile(data)

}
