% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{methylSigTileTFBS}
\alias{methylSigTileTFBS}
\title{Obtain tiled methylation data by tiling (pooling) data that a particular transcription factor (TF) is predicted to bind.}
\usage{
methylSigTileTFBS(meth, tfbsInfo)
}
\arguments{
\item{meth}{A \code{methylSigData-class} object used to tile data.}

\item{tfbsInfo}{An object that contains transcription factor binding sites information.}
}
\value{
A \code{methylSigData-class} object.
}
\description{
Gives total number of methylated cytosines and total number of reads (coverage) for each TF. This is used prior to the \code{\link{methylSigCalc}} function when the user prefers to conduct tests to identify significant level of hypermethylation or hypomethylation across their binding sites.
}
\seealso{
\code{\link{getTFBSInfo}} and \code{\link{methylSigTile}}.
}
