% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{methylSigWeightedAutoCorr}
\alias{methylSigWeightedAutoCorr}
\title{Weighted auto correlation of methylation rates}
\usage{
methylSigWeightedAutoCorr(chr, start, methRates, coverage, lags = 2)
}
\arguments{
\item{chr}{A single string indicating chromosome from `methylSigData' or `methylSigDiff' objects.}

\item{start}{A vector of CpG site loci from `methylSigData' or `methylSigDiff' objects.}

\item{methRates}{A numeric vector of methylation rates at each CpG site.}

\item{coverage}{A numeric vector of coverage levels (number of reads) at each CpG site.}

\item{lags}{Autocorrelation lags. This argument can be a vector if you want to calculate multiiple lags. Default is 2.}
}
\value{
Auto correlation value or vector for the lags entered.
}
\description{
This funciton calculates the weighted auto correlation of methylation rates based on the coverage level at each CpG site.
}
\details{
The weight for the locus i and j is \code{coverage[i]*coverage[j]/(coverage[i]+coverage[j])}.
}
\examples{
data(methylSig)

lags = c(2,5,10,50)

autoCorr = methylSigWeightedAutoCorr(meth[,"chr"], meth[,"start"],
             meth[,"numCs1"]/meth[,"coverage1"], meth[,"coverage1"],
             lags=lags)
autoCorr
}

