% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{refGeneAnnotationPlot}
\alias{refGeneAnnotationPlot}
\title{RefGene annotation plot}
\usage{
refGeneAnnotationPlot(listFrom, main = "Plot", priority = c("cds",
  "promoter", "noncoding", "5'utr", "3'utr"))
}
\arguments{
\item{listFrom}{Object from fuction `refGeneAnnotation'.}

\item{main}{An overall title for the plot. Default is "Plot".}

\item{priority}{A character vector indicating the priority order for assigning the annotated CpG site when overlapping occurs.  Default is coding sequence exon, promoter, noncoding exon, 5' UTR exon, 3' UTR exon, intron and then integenic region.  Only "cds", "promoter", "noncoding", "5'utr" and "3'utr" are acceptable.}
}
\description{
This function generates a refGene annotation bar chart.
}
\details{
Promoter is defined as the region 1,000bp upstream from a transcription start site (TSS) to the TSS. When regions overlap, the priority to assign regions for the CpG sites is defined by `priority' argument.
}
\examples{
data(methylSig)

refGeneInfo = getRefgeneInfo(system.file("annotation", "refGene.txt",
                 package = "methylSig"))

refGeneAnn = refGeneAnnotation(refGeneInfo, myDiffSigboth)

refGeneAnnotationPlot(refGeneAnn,main="ALL", priority=c("promoter","cds",
                           "noncoding", "5'utr", "3'utr"))
}
\seealso{
\code{\link{refGeneAnnotation}} and \code{\link{getRefgeneInfo}}
}
\keyword{refGeneAnnotation}

