% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylSig.R
\name{write.methylSigDiff}
\alias{write.methylSigDiff}
\title{Write a methylSigDiff object as a text file}
\usage{
write.methylSigDiff(object, ...)
}
\arguments{
\item{object}{a \code{methylSigDiff-class} object.}

\item{...}{Arguments inherited from \code{\link[utils]{write.table}}}
}
\value{
None
}
\description{
This funciton writes \code{methylSigDiff-class} as a text file. All options for write.table are avaiable for this function.
}
\examples{
data(sampleData)
write.methylSigDiff(myDiffSigboth, file="myResult.txt", row.names=FALSE, quote=FALSE, sep="\\t")
}

